requirements.txt
Flask==1.1.2
gunicorn==20.1.0
Pillow==9.0.0
numpy==1.22.0
matplotlib==3.5.1
tensorflow==2.7.0